requirements.txt
fastapi
uvicorn
transformers
torch
faiss-cpu
python-docx
docling-core
tqdm
numpy
Pillow
python-doctr[torch]
docling==2.42.0
python-multipart
openai>=1.3.0




